Another working demo but without react state and using MST observer in child component "sort-work-document-area"
